+++ rocket-fit-front/client/src/components/CustomizeWorkoutPages/CustomizeWorkoutTable.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import styled from "styled-components";
 import NumberAdjuster from "../NumberAdjuster";
 import AddExerciseModal from "./AddExerciseModal";
