+++ rocket-fit-front/client/src/components/MainPageContent/WeekButtons.tsx
@@ -1,4 +1,3 @@
-import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
