+++ rocket-fit-front/client/src/pages/AuthenticationPage.tsx
@@ -2,40 +2,9 @@ import { KeyboardEvent, useState } from "react";
 import authUserService from "../services/authUserService";
 import styled from "styled-components";
 import { Link, useNavigate } from "react-router-dom";
+import AuthenticationInput from "../components/AuthenticationInput";
 
 const LoginHeader1 = styled.h1``;
-
-const LoginHeader2 = styled.h2`
-  margin: 1em 0em 0em 0em;
-`;
-
-const LoginInput = styled.input`
-  margin: 0em 0em 1em 0em;
-`;
-
-const SubmitButton = styled.button`
-  margin: 1em 0em 0em 0em;
-  display: flex;
-  flex-direction: column;
-`;
-
-const RegisterButton = styled.button`
-  margin: 1em 0em;
-  text-decoration: none;
-`;
-
-const ButtonLink = styled(Link)`
-  text-decoration: none;
-  color: black;
-`;
-
-const SpinnerDiv = styled.div`
-  margin: 1em 0em 0em 0em;
-  display: flex;
-  flex-direction: column;
-`;
-
-const SpinnerSpan = styled.span``;
 
 const ErrorMessage = styled.div`
   background-color: red;
@@ -48,52 +17,16 @@ interface Props {
 }
 
 const AuthenticationPage = ({ sendDataToParent }: Props) => {
-  const navigate = useNavigate();
-  const [authId, setAuthId] = useState(-10);
-  const [username, setUsername] = useState<string>("");
-  const [password, setPassword] = useState<string>("");
   const [invalidLogin, setInvalidLogin] = useState<boolean>(false);
-  const [loading, setLoading] = useState<boolean>(false);
 
-  const handleInputChange = (
-    event: React.ChangeEvent<HTMLInputElement>,
-    inputVariable: string
-  ) => {
-    if (inputVariable === "username") {
-      const textValue: string = (event.target as HTMLInputElement).value;
-      setUsername(textValue);
-    } else if (inputVariable === "password") {
-      const textValue: string = (event.target as HTMLInputElement).value;
-      setPassword(textValue);
+  const handleInputData = async (returned_id: number) => {
+    if (returned_id === -10) {
+      setInvalidLogin(true);
+      sendDataToParent(-10);
+    } else {
+      setInvalidLogin(false);
+      sendDataToParent(returned_id);
     }
-  };
-
-  // Allows user to press enter to submit login
-  const handleEnterDown = (event: KeyboardEvent) => {
-    if (event.key === "Enter") {
-      SubmitLogin();
-    }
-  };
-
-  const SubmitLogin = () => {
-    setLoading(true);
-    const { request } = authUserService.getAll(
-      "/login?loginKey=" + username + "&password=" + password
-    );
-    request
-      .then((response) => {
-        const returned_id: number = response.data as unknown as number;
-        console.log(returned_id, authId);
-        if (returned_id === -10 && authId === -10) {
-          setInvalidLogin(true);
-          setLoading(false);
-        } else {
-          setAuthId(returned_id);
-          sendDataToParent(returned_id);
-          navigate("/myworkouts");
-        }
-      })
-      .catch((err) => console.log(err));
   };
 
   return (
@@ -104,38 +37,7 @@ const AuthenticationPage = ({ sendDataToParent }: Props) => {
       ) : (
         <EmptyDiv></EmptyDiv>
       )}
-      <LoginHeader2>Username:</LoginHeader2>
-      <LoginInput
-        key={1}
-        type="text"
-        placeholder=""
-        value={username}
-        onChange={(e) => handleInputChange(e, "username")}
-        onKeyDown={handleEnterDown}
-      />
-      <LoginHeader2>Password:</LoginHeader2>
-      <LoginInput
-        key={2}
-        type="password"
-        placeholder=""
-        value={password}
-        onChange={(e) => handleInputChange(e, "password")}
-        onKeyDown={handleEnterDown}
-      />
-      {loading ? (
-        <SpinnerDiv className="spinner-border text-dark" role="status">
-          <SpinnerSpan className="sr-only"></SpinnerSpan>
-        </SpinnerDiv>
-      ) : (
-        <>
-          <SubmitButton onClick={SubmitLogin} onKeyDown={handleEnterDown}>
-            Submit
-          </SubmitButton>
-          <RegisterButton>
-            <ButtonLink to={"/register"}>New User? Register Here</ButtonLink>
-          </RegisterButton>
-        </>
-      )}
+      <AuthenticationInput sendDataToPage={handleInputData} />
     </>
   );
 };
