+++ rocket-fit-front/client/src/pages/UnauthorizedPage.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
