+++ rocket-fit-front/client/src/pages/WorkoutsTemplatesPage.tsx
@@ -9,20 +9,10 @@ import exerciseService, { Exercise } from "../services/exerciseService";
 import { Link, useLocation, useNavigate } from "react-router-dom";
 import Spinner from "../components/Spinner";
 import DropdownTable from "../components/DropdownTable";
-
-const HeaderOne = styled.h1``;
-
-const StyledTable = styled.table``;
-
-const FlipButton = styled.button``;
+import Templates from "../components/WorkoutTemplateComponents/Templates";
+import CreateFromScratch from "../components/WorkoutTemplateComponents/CreateFromScratch";
 
 const BackButton = styled.button``;
-
-const CreateNewWorkoutDiv = styled.div`
-  padding: 3em 0em;
-`;
-
-const CreateNewWorkoutButton = styled.button``;
 
 interface Props {
   authId: number;
@@ -30,26 +20,11 @@ interface Props {
 
 const WorkoutTemplatesPage = ({ authId }: Props) => {
   const Navigate = useNavigate();
-
   //number of workouts the user already has
   const location = useLocation();
   const numberOfWorkouts = location.state;
-  console.log(numberOfWorkouts);
-
-  const [templates, setTemplates] = useState<StandardizedWorkoutTemplate[]>([]);
   const [exercises, setExercises] = useState<Exercise[]>([]);
-  const [dropdowns, setDropdowns] = useState<boolean[]>([]);
-  const [trigger, setTrigger] = useState<boolean>(false);
   const [loading, setLoading] = useState<boolean>(true);
-  const emptyWorkout = {
-    workoutName: "New Workout " + (numberOfWorkouts + 1),
-    weeks: 4,
-    days: [1],
-    exercises: [1],
-    sets: [3],
-    reps: [12],
-    rest: [120],
-  };
 
   if (authId == -10) {
     Navigate("/unauthorized");
@@ -65,60 +40,28 @@ const WorkoutTemplatesPage = ({ authId }: Props) => {
       .catch((err) => console.log(err));
   }, []);
 
-  useEffect(() => {
-    const { request } = workoutTemplateService.getAll("");
-    request
-      .then((response) => {
-        const workoutTemplates =
-          response.data as unknown[] as WorkoutTemplate[];
-        const standardTemplatesAndDropdowns =
-          standardizeWorkoutTemplates(workoutTemplates);
-        setTemplates(standardTemplatesAndDropdowns[0]);
-        setDropdowns(standardTemplatesAndDropdowns[1]);
-        setLoading(false);
-      })
-      .catch((err) => console.log(err));
-  }, []);
-
-  const FlipDrop = (count: number) => {
-    setTrigger(!trigger);
-    dropdowns[count] = !dropdowns[count];
-  };
+  useEffect(() => {}, [loading]);
 
   return (
     <>
       <BackButton>
         <Link to="/myworkouts">Back</Link>
       </BackButton>
+      <Templates
+        exercises={exercises}
+        numberOfWorkouts={numberOfWorkouts}
+        sendDataToParent={(loading) => setLoading(loading)}
+      />
       {loading ? (
-        <Spinner />
+        <>
+          <Spinner />
+        </>
       ) : (
         <>
-          {templates.map((item, count) => (
-            <>
-              <HeaderOne>{item.workoutName}</HeaderOne>
-              <FlipButton onClick={() => FlipDrop(count)}>flip</FlipButton>
-              {dropdowns[count] ? (
-                <DropdownTable
-                  item={item}
-                  exercises={exercises}
-                  numberOfWorkouts={numberOfWorkouts}
-                />
-              ) : (
-                <StyledTable></StyledTable>
-              )}
-            </>
-          ))}
-          <CreateNewWorkoutDiv>
-            <CreateNewWorkoutButton>
-              <Link
-                to="/customize"
-                state={[emptyWorkout, exercises, numberOfWorkouts]}
-              >
-                Create From Scratch
-              </Link>
-            </CreateNewWorkoutButton>
-          </CreateNewWorkoutDiv>
+          <CreateFromScratch
+            exercises={exercises}
+            numberOfWorkouts={numberOfWorkouts}
+          />
         </>
       )}
     </>
